src/requirements.txt
line-bot-sdk
flask
pytz
lxml
bs4
requests
python-dotenv
folium